#11 Created a Cable Tension Script
Cable Tension script is attached to the empty parent object that holds all the capsule that make up the whole cable. Each child is added to a list.

Next, we will try and set a maximum distance using Vector3 to prevent individual capsules from spazzing out when we pull the cable while grabbing.